update discord.js to v13 and refactor code
update discord.js to v13 (all packages now updated to latest version) and refactor as necessary, add package-lock.json to gitignore, fix ordering in command description, refactor slashCmdResp, use strings in saveddata for ids, correct spelling of ephemeral in commands and formats.txt, partial conversion of msg.channel.send to common.regCmdResp, rename some files, change names in common and common.handlers, update docker sh files, update Dockerfile, update formats.txt, clean up index, add package updating script

+++ commands/troll.js
@@ -9,36 +9,36 @@ module.exports = [
       if (rawArgs[1]) {
         if (rawArgs[0] == 'roles') {
           let promises = [ msg.channel.send('@ everyone').then(x => x.edit('@everyone')) ];
-          let roles = msg.guild.roles.cache.array().map(x => `<@&${x.id}>`);
+          let roles = Array.from(msg.guild.roles.cache.values()).map(x => `<@&${x.id}>`);
           for (var i = 0; i < roles.length; i += 90) {
-            promises.push(msg.channel.send(roles.slice(i, i + 90).join('')).then(x => x.delete()));
+            promises.push(msg.channel.send(roles.slice(i, i + 90).join(''), { allowedMentions: { parse: ['everyone'] } }).then(x => x.delete()));
           }
           return Promise.allSettled(promises);
         } else if (rawArgs[0] == 'members') {
           if (msg.guild.memberCount > 1000)
             return msg.channel.send('Error: too many members in guild to ping all members.');
           let promises = [ msg.channel.send('@ everyone').then(x => x.edit('@everyone')) ];
-          let members = msg.guild.members.cache.array().filter(x => !x.user.bot).map(x => `<@${x.id}>`);
+          let members = Array.from(msg.guild.members.cache.values()).filter(x => !x.user.bot).map(x => `<@${x.id}>`);
           for (var i = 0; i < members.length; i += 95) {
-            promises.push(msg.channel.send(members.slice(i, i + 95).join('')).then(x => x.delete()));
+            promises.push(msg.channel.send(members.slice(i, i + 95).join(''), { allowedMentions: { parse: ['everyone'] } }).then(x => x.delete()));
           }
           return Promise.allSettled(promises);
         }
       } else {
         if (rawArgs[0] == 'roles') {
           let promises = [];
-          let roles = msg.guild.roles.cache.array().map(x => `<@&${x.id}>`);
+          let roles = Array.from(msg.guild.roles.cache.values()).map(x => `<@&${x.id}>`);
           for (var i = 0; i < roles.length; i += 90) {
-            promises.push(msg.channel.send(roles.slice(i, i + 90).join('')));
+            promises.push(msg.channel.send(roles.slice(i, i + 90).join(''), { allowedMentions: { parse: ['everyone'] } }));
           }
           return Promise.allSettled(promises);
         } else if (rawArgs[0] == 'members') {
           if (msg.guild.memberCount > 1000)
             return msg.channel.send('Error: too many members in guild to ping all members.');
           let promises = [];
-          let members = msg.guild.members.cache.array().filter(x => !x.user.bot).map(x => `<@${x.id}>`);
+          let members = Array.from(msg.guild.members.cache.values()).filter(x => !x.user.bot).map(x => `<@${x.id}>`);
           for (var i = 0; i < members.length; i += 95) {
-            promises.push(msg.channel.send(members.slice(i, i + 95).join('')));
+            promises.push(msg.channel.send(members.slice(i, i + 95).join(''), { allowedMentions: { parse: ['everyone'] } }));
           }
           return Promise.allSettled(promises);
         }
@@ -51,9 +51,9 @@ module.exports = [
     flags: 0b010110,
     execute(o, msg, rawArgs) {
       if (!(common.isAdmin(msg) || msg.guild.id == persData.ids.guild.v0)) return;
-      let members = msg.guild.members.cache.array().filter(x => !x.user.bot);
+      let members = Array.from(msg.guild.members.cache.values()).filter(x => !x.user.bot);
       let random_member = members[common.randInt(0, members.length)];
-      return msg.channel.send(`Random ping: <@!${random_member.user.id}>`);
+      return msg.channel.send(`Random ping: <@!${random_member.user.id}>`, { allowedMentions: { parse: ['everyone'] } });
     },
   },
 ];
